fix silly bugs

+++ src/main/scala/org/loopring/lightcone/lib/ring/RingSerializer.scala
@@ -144,11 +144,11 @@ private[lib] class RingSerializerHelper(lrcAddress: String, ring: Ring) {
     insertOffset(datastream.addUint32(order.validSince, false))
 
     orderSpendableSMap.get(order.hash) match {
-      case Some(x: Int) ⇒ tablestream.addUint16(x.intValue())
+      case Some(x: Int) ⇒ tablestream.addUint16(x)
       case _            ⇒ throw new Exception("ringGenerator get " + order.hash + "orderSpendableS failed")
     }
     orderSpendableFeeMap.get(order.hash) match {
-      case Some(x: Int) ⇒ tablestream.addUint16(x.intValue())
+      case Some(x: Int) ⇒ tablestream.addUint16(x)
       case _            ⇒ throw new Exception("ringGenerator get " + order.hash + "orderSpendableFee failed")
     }
